Cleanup

+++ benchmark/src/benchmark_util.h
@@ -1,6 +1,6 @@
 #include <string>
 
-static inline std::string gen_random(const int len)
+static inline std::string gen_random(const size_t len)
 {
   static const char alphanum[] =
       "0123456789"
@@ -9,9 +9,9 @@ static inline std::string gen_random(const int len)
   std::string tmp_s;
   tmp_s.reserve(len);
 
-  for (int i = 0; i < len; ++i)
+  for (size_t i = 0; i < len; ++i)
   {
-    tmp_s += alphanum[rand() % (sizeof(alphanum) - 1)];
+    tmp_s += alphanum[static_cast<size_t>(rand()) % (sizeof(alphanum) - 1)];
   }
 
   return tmp_s;
